# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Tailor CRM API
  version: 3.0.0
  description: |
    # Production-Ready REST API for Tailor Management SaaS

    **Last Updated:** May 10, 2026
    **Status:** Fully Implemented & Tested
    **Total Endpoints:** 80 (73 + 7 system)

    ## Response Format (ALL endpoints)

    Success:   { success: true,  message: "...", data: {} }
    Error:     { success: false, message: "...", error: { code: "...", details: {} } }
    Paginated: { success: true,  message: "...", data: { items: [], pagination: { page, limit, total } } }


    ## Field Naming Convention
    - Requests:  snake_case  (customer_id, delivery_date)
    - Responses: camelCase   (customerId, deliveryDate)

    ## Order Status Flow
    pending → in_progress → ready → delivered  (cannot skip)

    ## Error Codes
    INVALID_INPUT | UNAUTHORIZED | FORBIDDEN | NOT_FOUND
    DUPLICATE_PHONE | INVALID_STATUS_TRANSITION
    CUSTOMER_NOT_FOUND | ORDER_NOT_FOUND | SHOP_NOT_FOUND
    MEASUREMENT_NOT_FOUND | STAFF_NOT_FOUND | INVOICE_NOT_FOUND
    INTERNAL_ERROR | RATE_LIMIT_EXCEEDED | SERVICE_UNAVAILABLE

    ## Rate Limiting
    All requests are rate-limited per IP address:
    - **Global**: 100 requests / 15 minutes (default)
    - **Login**: 5 requests / 15 minutes
    - **OTP/Verify**: 3 requests / 15 minutes
    - **Refresh**: 10 requests / 15 minutes
    - **Upload**: 20 requests / 15 minutes
    
    Rate limit status in response headers:
    - `RateLimit-Limit`: Maximum requests allowed
...

...

...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        ## JWT Authentication with Token Rotation & Blacklist
        
        **Obtain tokens from:**
        - POST /auth/login (Firebase production)
        - POST /auth/login-test (Development only)
        
        **Token Structure:**
        - Access Token: 7-day expiry, contains userId, phone, shopId, sessionId, jti
        - Refresh Token: Long-lived, used for rotation
        
        **Token Rotation Behavior:**
        - Every refresh returns NEW access + refresh tokens
        - Old refresh token becomes invalid (single-use)
        - Prevents token reuse attacks
        - Session ID maintained across rotations
        
        **Security Features:**
        - Redis-backed session tracking
        - Access token blacklist on logout
        - Session activity validation
        - JTI (JWT ID) claim prevents replay attacks
        - Automatic revocation on suspicious activity
        
        **On 401 Unauthorized:**
        - Token expired? Call POST /auth/refresh-token
        - Still fails? Re-login with POST /auth/login
        - Token revoked? Session blacklisted, must re-login
        
        **Usage:**
        Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
...
security:
  - BearerAuth: []
...
paths:
...
  /auth/verify-token:
    post:
      tags: [Authentication]
      summary: Verify JWT Token validity
      operationId: verifyToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
            example:
              token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOjF9.sig"
      responses:
        '200':
          description: Token is valid
          content:
            application/json:
              example:
                success: true
                message: Token is valid
                data:
                  userId: 1
                  phone: "+919705116606"
                  shopId: 1
                  isValid: true
                  expiresAt: "2026-05-06T10:00:00Z"
        '401':
          description: Token invalid or expired
          content:
            application/json:
              example:
                success: false
                message: Invalid or expired token
                error:
                  code: UNAUTHORIZED
  /auth/refresh-token:
    post:
      tags: [Authentication]
      summary: Refresh JWT Token
      description: Use refresh token to get new access token. Both tokens are rotated.
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
            example:
              refreshToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOjF9.ref"
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              example:
                success: true
                message: Token refreshed successfully
                data:
                  token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOjF9.newSig"
                  refreshToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOjF9.newRef"
                  expiresIn: 604800
        '401':
          description: Invalid refresh token
          content:
            application/json:
              example:
                success: false
                message: Invalid or expired refresh token
                error:
                  code: UNAUTHORIZED
  /auth/profile:
    get:
      tags: [Authentication]
      summary: Get current user profile
      operationId: getProfile
      responses:
        '200':
          description: Profile retrieved
          content:
            application/json:
              example:
                success: true
                message: Profile retrieved successfully
                data:
                  id: 1
                  phone: "+919705116606"
                  name: "Siddiq Khan"
                  shopId: 1
                  createdAt: "2026-04-01T10:00:00Z"
    put:
      tags: [Authentication]
      summary: Update user profile
      operationId: updateProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
            example:
              name: "Siddiq Mohammed Khan"
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              example:
                success: true
                message: Profile updated successfully
                data:
                  id: 1
                  phone: "+919705116606"
                  name: "Siddiq Mohammed Khan"
                  shopId: 1
...
  /activity/order/{orderId}:
    get:
      tags: [Activity]
      summary: Get activity timeline for an order
      operationId: getActivityByOrder
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Activity logs retrieved
          content:
            application/json:
              example:
                success: true
                message: Activity logs retrieved successfully
                data:
                  items:
                    - id: 3
                      orderId: 1
                      actionType: "comment"
                      notes: "Customer called, needs delivery by Friday"
                      createdAt: "2026-04-29T15:00:00Z"
                    - id: 2
                      orderId: 1
                      actionType: "payment"
                      newValue: "500.00"
                      notes: "Payment recorded: ₹500 via upi"
                      createdAt: "2026-04-29T12:00:00Z"
                    - id: 1
                      orderId: 1
                      actionType: "status_change"
                      oldValue: "pending"
                      newValue: "in_progress"
                      notes: "Status changed from pending to in_progress"
                      createdAt: "2026-04-29T11:00:00Z"

    post:
      tags: [Activity]
      summary: Add comment/note to order activity log
      operationId: addActivityNote
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [notes]
              properties:
                notes:
                  type: string
            example:
              notes: "Customer called, needs delivery by Friday urgently"
      responses:
        '201':
          description: Comment added
          content:
            application/json:
              example:
                success: true
                message: Activity note added successfully
                data:
                  id: 4
                  orderId: 1
                  actionType: "comment"
                  notes: "Customer called, needs delivery by Friday urgently"
                  createdAt: "2026-04-29T16:00:00Z"

  # ══════════════════════════════════════════════════════
  # DASHBOARD — 1 endpoint
  # ══════════════════════════════════════════════════════
...